+++ shake/compiler/parser/test-generator/index.js
@@ -1,0 +1,2 @@
+require("./lib/class-tests.js");
+require("./lib/field-tests.js");