+++ shake/compiler/shaketest/webpack.config.d/resolve.js
@@ -1,0 +1,5 @@
+config.resolve.alias = {
+    "crypto": false,
+    "path": require.resolve("path-browserify"),
+    "fs": false,
+}